requirements.txt
fastapi==0.112.0
uvicorn==0.30.3
pdfminer.six==20231228
requests==2.32.3
pydantic==1.10.15
